NODE-1283: Share as much code as possible between Block/Deploy download managers
