+++ comm/src/main/scala/io/casperlabs/comm/gossiping/downloadmanager/DownloadManager.scala
@@ -1,17 +1,26 @@
 package io.casperlabs.comm.gossiping.downloadmanager
 
 import cats._
-import cats.effect.concurrent.Deferred
+import cats.effect._
+import cats.effect.concurrent._
+import cats.effect.implicits._
 import cats.implicits._
+import com.google.protobuf.ByteString
 import eu.timepit.refined._
 import eu.timepit.refined.api.Refined
 import eu.timepit.refined.auto._
 import eu.timepit.refined.numeric._
+import io.casperlabs.comm.GossipError
 import io.casperlabs.comm.discovery.Node
-import io.casperlabs.comm.gossiping.WaitHandle
+import io.casperlabs.comm.discovery.NodeUtils.showNode
+import io.casperlabs.comm.gossiping.{Chunk, GossipService, Relaying, WaitHandle}
 import io.casperlabs.metrics.Metrics
+import io.casperlabs.shared.Log._
+import io.casperlabs.shared.{Compression, Log}
+import monix.tail.Iterant
 
 import scala.concurrent.duration.{Duration, FiniteDuration}
+import scala.util.control.NonFatal
 
 trait DownloadManagerTypes {
 
@@ -35,26 +44,15 @@ trait DownloadManagerTypes {
   type Downloadable
 }
 
-/** Manages the download, validation, storing and gossiping of [[DownloadManagerTypes#Downloadable]].*/
-trait DownloadManager[F[_]] extends DownloadManagerTypes {
-
-  /** Schedule the download of a full [[Downloadable]] from the `source` node by a [[Handle]].
-    * If `relay` is `true` then gossip it afterwards, if it's valid.
-    * The returned `F[F[Unit]]` represents the success/failure of
-    * the scheduling itself; it fails if there's any error accessing
-    * the local backend, or if the scheduling cannot be carried out
-    * due to missing dependencies (at this point we should have synced
-    * already and the schedule should be called in topological order).
-    *
-    * The unwrapped `F[Unit]` _inside_ the `F[F[Unit]]` can be used to
-    * wait until the actual download finishes, or results in an error. */
-  def scheduleDownload(handle: Handle, source: Node, relay: Boolean): F[WaitHandle[F]]
-}
-
 trait DownloadManagerCompanion extends DownloadManagerTypes {
 
+  implicit val metricsSource: Metrics.Source
+
+  // to use .minimumOption on collections of (Node, Int)
+  implicit val sourcesAndCountersOrder: Order[(Node, Int)] = Order[Int].contramap[(Node, Int)](_._2)
+
   /** Export base 0 values so we have non-empty series for charts. */
-  def establishMetrics[F[_]: Monad: Metrics](implicit ev: Metrics.Source) =
+  def establishMetrics[F[_]: Monad: Metrics] =
     for {
       _ <- Metrics[F].incrementCounter("downloads_failed", 0)
       _ <- Metrics[F].incrementCounter("downloads_succeeded", 0)
@@ -124,4 +122,421 @@ trait DownloadManagerCompanion extends DownloadManagerTypes {
 
   /** All dependencies that need to be downloaded before a downloadable. */
   def dependencies(handle: Handle): Seq[Identifier]
+}
+
+/** Manages the download, validation, storing and gossiping of [[DownloadManagerTypes#Downloadable]].*/
+trait DownloadManager[F[_]] extends DownloadManagerTypes {
+
+  /** Schedule the download of a full [[Downloadable]] from the `source` node by a [[Handle]].
+    * If `relay` is `true` then gossip it afterwards, if it's valid.
+    * The returned `F[F[Unit]]` represents the success/failure of
+    * the scheduling itself; it fails if there's any error accessing
+    * the local backend, or if the scheduling cannot be carried out
+    * due to missing dependencies (at this point we should have synced
+    * already and the schedule should be called in topological order).
+    *
+    * The unwrapped `F[Unit]` _inside_ the `F[F[Unit]]` can be used to
+    * wait until the actual download finishes, or results in an error. */
+  def scheduleDownload(handle: Handle, source: Node, relay: Boolean): F[WaitHandle[F]]
+}
+
+trait DownloadManagerImpl[F[_]] extends DownloadManager[F] { self =>
+  // Workaround to reuse the existing code
+  val companion: DownloadManagerCompanion
+
+  override type Handle       = companion.Handle
+  override type Identifier   = companion.Identifier
+  override type Downloadable = companion.Downloadable
+
+  import companion._
+
+  // Can't specify context bounds in traits
+  implicit val C: Concurrent[F]
+  implicit val T: Timer[F]
+  implicit val L: Log[F]
+  implicit val M: Metrics[F]
+  implicit val E: Eq[Identifier]
+  implicit val S: Show[Identifier]
+  // For graceful shutdown
+  val isShutdown: Ref[F, Boolean]
+  // Keep track of active downloads and dependencies.
+  val itemsRef: Ref[F, Map[Identifier, companion.Item[F]]]
+  // Keep track of ongoing downloads so we can cancel them.
+  val workersRef: Ref[F, Map[Identifier, Fiber[F, Unit]]]
+  // Limit parallel downloads.
+  val semaphore: Semaphore[F]
+  // Single item control signals for the manager loop.
+  val signal: MVar[F, companion.Signal[F]]
+  // Establish gRPC connection to another node.
+  val connectToGossip: GossipService.Connector[F]
+  val backend: companion.Backend[F]
+  val relaying: Relaying[F]
+  val retriesConf: companion.RetriesConf
+
+  def toByteString(id: Identifier): ByteString
+
+  implicit class IdentifierOps(id: Identifier) {
+    def toByteString: ByteString = self.toByteString(id)
+  }
+
+  // 'From*' because of a compilation error `have same type after erasure: (handle: Object)Object`
+  def extractIdFromHandle(handle: Handle): Identifier
+  def extractIdFromDownloadable(downloadable: Downloadable): Identifier
+
+  implicit class HandleOps(h: Handle) {
+    def id: Identifier = extractIdFromHandle(h)
+  }
+
+  implicit class DownloadableOps(d: Downloadable) {
+    def id: Identifier = extractIdFromDownloadable(d)
+  }
+  def streamChunks(source: Node, id: Identifier): Iterant[F, Chunk]
+  def parseDownloadable(bytes: Array[Byte]): Downloadable
+
+  private def ensureNotShutdown: F[Unit] =
+    isShutdown.get.ifM(
+      Sync[F]
+        .raiseError(new java.lang.IllegalStateException("Download Manager already shut down.")),
+      Sync[F].unit
+    )
+
+  private def makeDownloadFeedback =
+    Deferred[F, Either[Throwable, Unit]]
+
+  private def makeScheduleFeedback =
+    Deferred[F, Either[Throwable, Deferred[F, Either[Throwable, Unit]]]]
+
+  /** Schedule the download of a full [[Downloadable]] from the `source` node by a [[Handle]].
+    * If `relay` is `true` then gossip it afterwards, if it's valid.
+    * The returned `F[F[Unit]]` represents the success/failure of
+    * the scheduling itself; it fails if there's any error accessing
+    * the local backend, or if the scheduling cannot be carried out
+    * due to missing dependencies (at this point we should have synced
+    * already and the schedule should be called in topological order).
+    *
+    * The unwrapped `F[Unit]` _inside_ the `F[F[Unit]]` can be used to
+    * wait until the actual download finishes, or results in an error. */
+  def scheduleDownload(handle: Handle, source: Node, relay: Boolean): F[WaitHandle[F]] =
+    for {
+      // Fail rather than block forever.
+      _ <- ensureNotShutdown
+      // Feedback about whether we successfully scheduled the item.
+      sf <- makeScheduleFeedback
+      _  <- signal.put(Signal.Download(handle, source, relay, sf))
+      df <- Sync[F].rethrow(sf.get)
+    } yield Sync[F].rethrow(df.get)
+
+  /** Run the manager loop which listens to signals and starts workers when it can. */
+  def run: F[Unit] =
+    signal.take.flatMap {
+      case Signal.Download(handle, source, relay, scheduleFeedback) =>
+        // At this point we should have already synced and only scheduled things to which we know how to get.
+        val start =
+          isDownloaded(handle.id).ifM(
+            for {
+              downloadFeedback <- makeDownloadFeedback
+              _                <- downloadFeedback.complete(Right(()))
+            } yield downloadFeedback,
+            for {
+              _                        <- ensureNoMissingDependencies(handle)
+              items                    <- itemsRef.get
+              itemAndFeedback          <- mergeItem(items, handle, source, relay)
+              (item, downloadFeedback) = itemAndFeedback
+              _                        <- backend.onScheduled(handle).start.whenA(!items.contains(handle.id))
+              _                        <- itemsRef.update(_ + (handle.id -> item))
+              _                        <- if (item.canStart) startWorker(item) else Sync[F].unit
+              _                        <- setScheduledGauge()
+            } yield downloadFeedback
+          )
+        // Report any startup errors so the caller knows something's fatally wrong, then carry on.
+        start
+          .attemptAndLog("An error occurred when handling Download signal.")
+          .flatMap(scheduleFeedback.complete) >> run
+
+      case Signal.DownloadSuccess(id) =>
+        val finish = for {
+          _ <- workersRef.update(_ - id)
+          // Remove the item and check what else we can download now.
+          next <- itemsRef.modify { items =>
+                   val item = items(id)
+                   val dependants = items.collect {
+                     case (hash, dep) if dep.dependencies contains id =>
+                       hash -> dep.copy(dependencies = dep.dependencies - id)
+                   }
+                   val startables = dependants.collect {
+                     case (_, dep) if dep.canStart => dep
+                   }
+                   (items ++ dependants - id, item.maybeWatcher.toList -> startables.toList)
+                 }
+          (watchers, startables) = next
+          _                      <- watchers.traverse(_.complete(Right(())).attempt.void)
+          _                      <- backend.onDownloaded(id).start
+          _                      <- startables.traverse(startWorker)
+          _                      <- setScheduledGauge()
+        } yield ()
+
+        finish.attemptAndLog("An error occurred when handling DownloadSuccess signal.") >> run
+
+      case Signal.DownloadFailure(id, ex) =>
+        val finish = for {
+          _ <- workersRef.update(_ - id)
+          // Keep item so its dependencies are not downloaded.
+          // If it's scheduled again we'll try once more.
+          // Old stuff will be forgotten when the node is restarted.
+          watchers <- itemsRef.modify { items =>
+                       val item = items(id)
+                       val tombstone: Item[F] =
+                         item.copy(isDownloading = false, isError = true, maybeWatcher = None)
+                       (items + (id -> tombstone), item.maybeWatcher.toList)
+                     }
+          // Tell whoever scheduled it before that it's over.
+          _ <- watchers.traverse(_.complete(Left(ex)).attempt.void)
+          _ <- setScheduledGauge()
+        } yield ()
+
+        finish.attemptAndLog("An error occurred when handling DownloadFailure signal.") >> run
+    }
+
+  // Indicate how many items we have in the queue.
+  private def setScheduledGauge() =
+    for {
+      items <- itemsRef.get
+      _     <- Metrics[F].setGauge("downloads_scheduled", items.size.toLong)
+    } yield ()
+
+  /** Either create a new item or add the source to an existing one.
+    * Return the deferred that will eventually signal the completion of the task.
+    */
+  private def mergeItem(
+      items: Map[Identifier, Item[F]],
+      handle: Handle,
+      source: Node,
+      relay: Boolean
+  ): F[(Item[F], DownloadFeedback[F])] =
+    makeDownloadFeedback flatMap { freshDownloadFeedback =>
+      items.get(handle.id) map { existing =>
+        Sync[F].pure {
+          // If it was already scheduled we keep the existing Deferred.
+          val downloadFeedback = existing.maybeWatcher getOrElse freshDownloadFeedback
+          existing.copy(
+            sources = existing.sources + source,
+            relay = existing.relay || relay,
+            maybeWatcher = Some(downloadFeedback)
+          ) -> downloadFeedback
+        }
+      } getOrElse {
+        // Collect which dependencies have already been downloaded.
+        dependencies(handle).toList.traverse { hash =>
+          if (items.contains(hash)) Sync[F].pure(hash -> false)
+          else isDownloaded(hash).map(hash            -> _)
+        } map { deps =>
+          val pending = deps.filterNot(_._2).map(_._1).toSet
+          Item(
+            handle,
+            Set(source),
+            relay,
+            dependencies = pending,
+            maybeWatcher = Some(freshDownloadFeedback)
+          ) -> freshDownloadFeedback
+        }
+      }
+    }
+
+  /** Check that we either have all dependencies already downloaded or scheduled. */
+  private def ensureNoMissingDependencies(handle: Handle): F[Unit] =
+    dependencies(handle).toList.traverse { id =>
+      isScheduled(id).ifM((id -> true).pure[F], isDownloaded(id).map(id -> _))
+    } map {
+      _.filterNot(_._2).map(_._1)
+    } flatMap {
+      case Nil =>
+        Sync[F].unit
+      case missing =>
+        Sync[F].raiseError(
+          GossipError.MissingDependencies(handle.id.toByteString, missing.map(_.toByteString))
+        )
+    }
+
+  private def isScheduled(id: Identifier): F[Boolean] =
+    itemsRef.get.map(_ contains id)
+
+  private def isDownloaded(id: Identifier): F[Boolean] =
+    backend.contains(id)
+
+  /** Kick off the download and mark the item. */
+  private def startWorker(item: Item[F]): F[Unit] =
+    for {
+      _ <- itemsRef.update(_ + (item.handle.id -> item.copy(isDownloading = true)))
+      worker <- Concurrent[F].start {
+                 // Indicate how many items are currently being attempted, including their retry wait time.
+                 Metrics[F].gauge("downloads_ongoing") {
+                   download(item.handle.id)
+                 }
+               }
+      _ <- workersRef.update(_ + (item.handle.id -> worker))
+    } yield ()
+
+  // Just say which identifier to download, try all possible sources.
+  private def download(id: Identifier): F[Unit] = {
+    val success                = signal.put(Signal.DownloadSuccess(id))
+    def failure(ex: Throwable) = signal.put(Signal.DownloadFailure(id, ex))
+
+    def tryDownload(handle: Handle, source: Node, relay: Boolean) =
+      for {
+        downloadable <- fetchAndRestore(source, id)
+        _            <- backend.validate(downloadable)
+        _            <- backend.store(downloadable)
+        _            <- relaying.relay(List(handle.id.toByteString)).whenA(relay)
+        _            <- success
+        _            <- Metrics[F].incrementCounter("downloads_succeeded")
+      } yield ()
+
+    // Try to download until we succeed or give up.
+    def loop(counterPerSource: Map[Node, Int], lastError: Option[Throwable]): F[Unit] =
+      itemsRef.get.map(_(id)).flatMap { item =>
+        val prevSources           = counterPerSource.keySet
+        val potentiallyNewSources = item.sources
+        if (potentiallyNewSources.size > prevSources.size) {
+          val newSources = potentiallyNewSources.map((_, 0)).toMap |+| counterPerSource
+          // We've got new sources during download, so need to recheck the least tried source
+          loop(newSources, lastError)
+        } else {
+          // No sources changes, can continue.
+          // Finding least tried source.
+          val idAsString = id.show
+          counterPerSource.toList.minimumOption match {
+            case None =>
+              failure(new IllegalStateException("No source to download from."))
+
+            case Some((_, attempt)) if attempt > retriesConf.maxRetries.toInt =>
+              // Least tried source is tried more than allowed => don't have sources to try anymore.
+              // Let's just return the last error, unwrapped, so callers don't have to anticipate
+              // whether this component is going to do retries or not.
+              // Alternatively we could use `Throwable.addSupressed` to collect all of them.
+              val countersSum = counterPerSource.values.sum
+              val sourcesNum  = counterPerSource.size
+              val ex          = lastError.getOrElse(new IllegalStateException("No errors captured."))
+              Log[F]
+                .error(
+                  s"Could not download block ${idAsString} from any of the sources; tried $countersSum requests for $sourcesNum sources: $ex"
+                ) *> failure(ex)
+
+            case Some((source, attempt)) =>
+              val duration = if (attempt > 0) {
+                retriesConf.initialBackoffPeriod *
+                  math.pow(retriesConf.backoffFactor, (attempt - 1).toDouble)
+              } else {
+                Duration.Zero
+              }
+              duration match {
+                case delay: FiniteDuration =>
+                  Log[F].debug(
+                    s"Scheduling download of block ${idAsString} from ${source.show -> "peer"} later, $attempt, $delay"
+                  ) *>
+                    Timer[F].sleep(delay) *>
+                    tryDownload(item.handle, source, item.relay).handleErrorWith {
+                      case NonFatal(ex) =>
+                        val nextAttempt = attempt + 1
+                        Metrics[F].incrementCounter("downloads_failed") *>
+                          Log[F].warn(
+                            s"Retrying download of block ${idAsString} from other sources, failed source: ${source.show -> "peer"}, prev $attempt: $ex"
+                          ) *>
+                          loop(counterPerSource.updated(source, nextAttempt), ex.some)
+                    }
+                case _: Duration.Infinite => sys.error("Unreachable")
+              }
+          }
+        }
+      }
+
+    itemsRef.get.map(_(id)).flatMap { item =>
+      // Make sure the manager knows we're done, even if we fail unexpectedly.
+      loop(item.sources.map((_, 0)).toMap, lastError = none[Throwable]) recoverWith {
+        case NonFatal(ex) => failure(ex)
+      }
+    }
+  }
+
+  /** Download from the source node and decompress it. */
+  private def fetchAndRestore(source: Node, id: Identifier): F[Downloadable] = {
+    def invalid(msg: String) =
+      GossipError.InvalidChunks(msg, source)
+
+    // Keep track of how much we have downloaded so far and cancel the stream if it goes over the promised size.
+    case class Acc(
+        header: Option[Chunk.Header],
+        totalSizeSoFar: Int,
+        chunks: List[ByteString],
+        error: Option[GossipError]
+    ) {
+      def invalid(msg: String): Acc =
+        copy(error = Some(GossipError.InvalidChunks(msg, source)))
+
+      def append(data: ByteString): Acc =
+        copy(totalSizeSoFar = totalSizeSoFar + data.size, chunks = data :: chunks)
+    }
+
+    val effect =
+      for {
+        acc <- streamChunks(source, id).foldWhileLeftL(Acc(None, 0, Nil, None)) {
+                case (acc, chunk) if acc.header.isEmpty && chunk.content.isHeader =>
+                  val header = chunk.getHeader
+                  header.compressionAlgorithm match {
+                    case "" | "lz4" =>
+                      Left(acc.copy(header = Some(header)))
+                    case other =>
+                      Right(
+                        acc.invalid(s"Block chunks compressed with unexpected algorithm: $other")
+                      )
+                  }
+
+                case (acc, chunk) if acc.header.nonEmpty && chunk.content.isHeader =>
+                  Right(acc.invalid("Block chunks contained a second header."))
+
+                case (acc, _) if acc.header.isEmpty =>
+                  Right(acc.invalid("Block chunks did not start with a header."))
+
+                case (acc, chunk) if chunk.getData.isEmpty =>
+                  Right(acc.invalid("Block chunks contained empty data frame."))
+
+                case (acc, chunk)
+                    if acc.totalSizeSoFar + chunk.getData.size > acc.header.get.contentLength =>
+                  Right(acc.invalid("Block chunks are exceeding the promised content length."))
+
+                case (acc, chunk) =>
+                  Left(acc.append(chunk.getData))
+              }
+
+        content <- if (acc.error.nonEmpty) {
+                    Sync[F].raiseError[Array[Byte]](acc.error.get)
+                  } else if (acc.header.isEmpty) {
+                    Sync[F].raiseError[Array[Byte]](invalid("Did not receive a header."))
+                  } else {
+                    val header  = acc.header.get
+                    val content = acc.chunks.toArray.reverse.flatMap(_.toByteArray)
+                    if (header.compressionAlgorithm.isEmpty) {
+                      Sync[F].pure(content)
+                    } else {
+                      Compression
+                        .decompress(content, header.originalContentLength)
+                        .fold(
+                          Sync[F].raiseError[Array[Byte]](invalid("Could not decompress chunks."))
+                        )(Sync[F].pure(_))
+                    }
+                  }
+
+        downloadable <- Sync[F].delay(parseDownloadable(content))
+        _ <- Sync[F]
+              .raiseError(invalid("Retrieved block has unexpected block hash."))
+              .whenA(downloadable.id =!= id)
+      } yield downloadable
+
+    // Indicate how many fetches we are trying to do at a time. If it's larger then the semaphore
+    // we configured we'd know where we'd have to raise it to allow maximum throughput.
+    Metrics[F].gauge("fetches_ongoing") {
+      semaphore.withPermit {
+        effect
+      }
+    }
+  }
 }
